Edit Linters
