+++ src/index.html
@@ -69,7 +69,7 @@
                 </table>
             </div>
             <div class="clear-all">
-                <p class="clear-title">Clear all completed</p>
+                <p class="clear-title" id="clear-title">Clear all completed</p>
             </div>
         </div>
     </div>
